+++ expression/constant.go
@@ -53,7 +53,7 @@ func NewZero() *Constant {
 }
 
 // NewUInt64Const stands for constant of a given number.
-func NewUInt64Const(num int64) *Constant {
+func NewUInt64Const(num int) *Constant {
 	retT := types.NewFieldType(mysql.TypeLonglong)
 	retT.AddFlag(mysql.UnsignedFlag) // shrink range to avoid integral promotion
 	retT.SetFlen(mysql.MaxIntWidth)
